hm_term2_rest_api_projects/11a_04_23/requirements.txt
appdirs==1.4.3
cloudinary==1.7.0
decorator==4.0.11
dj-database-url==0.4.2
Django==1.8.4
django-cors-headers==2.0.2
django-pipeline==1.6.12
django-webpack-loader==0.2.4
djangorestframework==3.3.2
environ==1.0
gunicorn==19.7.1
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.2.0
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
urllib3==1.20
wcwidth==0.1.7
whitenoise==3.3.0
